add: helper for column names
